Enable HTML anchor in Details block
